examples/requirements-slack.txt
git+ssh://git@github.com/drydotai/structured-workspace.git@alpha
python-dotenv
slack-sdk
